+++ tests/templates/kuttl/oidc/login.py
@@ -23,7 +23,7 @@ keycloak_service = f"keycloak.{namespace}.svc.cluster.local"
 login_page = session.get(f"https://{druid_router_service}:9088/unified-console.html", verify=False, headers={'Content-type': 'application/json'})
 keycloak_base_url = f"https://{keycloak_service}:8443" if tls == 'true' else f"http://{keycloak_service}:8080"
 assert login_page.ok, "Redirection from Druid to Keycloak failed"
-assert login_page.url.startswith(f"{keycloak_base_url}/realms/test/protocol/openid-connect/auth?scope=openid+profile+email&response_type=code&redirect_uri=https%3A%2F%2F{druid_router_service}%3A9088%2Fdruid-ext%2Fdruid-pac4j%2Fcallback&state="), \
+assert login_page.url.startswith(f"{keycloak_base_url}/realms/test/protocol/openid-connect/auth?scope=openid+email+profile&response_type=code&redirect_uri=https%3A%2F%2F{druid_router_service}%3A9088%2Fdruid-ext%2Fdruid-pac4j%2Fcallback&state="), \
     "Redirection to Keycloak expected"
 
 # Login to keycloak with test user