+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 
 - Add support for specifying additional extensions to load ([#547], [#563]).
+- Add support for OIDC as authentication method ([#573]).
 
 ### Changed
 
@@ -14,8 +15,8 @@ All notable changes to this project will be documented in this file.
 - Bump `product-config` from `0.6.0` to `0.7.0` ([#585]).
 - Bump other dependencies ([#587]).
 
-[#585]: https://github.com/stackabletech/airflow-operator/pull/585
-[#587]: https://github.com/stackabletech/airflow-operator/pull/587
+[#585]: https://github.com/stackabletech/druid-operator/pull/585
+[#587]: https://github.com/stackabletech/druid-operator/pull/587
 
 ### Fixed
 
@@ -27,6 +28,7 @@ All notable changes to this project will be documented in this file.
 [#557]: https://github.com/stackabletech/druid-operator/pull/557
 [#563]: https://github.com/stackabletech/druid-operator/pull/563
 [#572]: https://github.com/stackabletech/druid-operator/pull/572
+[#573]: https://github.com/stackabletech/druid-operator/pull/573
 
 ## [24.3.0] - 2024-03-20